+++ src/App.tsx
@@ -1,6 +1,9 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { Outlet } from "react-router-dom";
+import { useRecoilState } from "recoil";
 import styled from "styled-components";
+import { getLoggedInInfo } from "./api/accountApi";
+import { ILoggedInAtom, loggedInAtom } from "./atoms/loggedInAtom";
 
 const Wrapper = styled.div`
   width: 100%;
@@ -8,6 +11,14 @@ const Wrapper = styled.div`
 `;
 
 function App() {
+  const [loggedIn, setLoggedIn] = useRecoilState<ILoggedInAtom>(loggedInAtom);
+  useEffect(() => {
+    getLoggedInInfo().then((res) => {
+      res.data === ""
+        ? setLoggedIn({ isLoggedIn: false, id: "" })
+        : setLoggedIn({ isLoggedIn: true, id: res.data });
+    });
+  }, []);
   return (
     <Wrapper>
       <Outlet />
